+++ modules/loader.py
@@ -7,9 +7,11 @@ import warnings
 import urllib3
 from modules import timer, errors
 
+
 initialized = False
 errors.install()
 logging.getLogger("DeepSpeed").disabled = True
+
 
 import torch # pylint: disable=C0411
 try: